# Use muladd for LSTM cell matmuls