This appears to help significantly on CPU, ref. https://discourse.julialang.org/t/slow-lstm-on-gpu-in-flux/84228. I have not yet benchmarked GPU performance, but it is likely a wash. It would be great to have a single function that does `x1 * y1 + x2 * y2`, does one exist in LinearAlgebra?