@mcabbott 2022-07-22
Great if `muladd` helps something, I think I tried for some other layers and was disappointed.

It is a very simple `out .= z; mul!(out, A, B)` which you could easily extend to allow `(A*B + C*D) .+ z`, but as far as I know nobody has packaged that up.