@mcabbott 2022-09-04
Timing as in this gist: https://gist.github.com/JLDC/66c8eed6c36cb9cda85ff6404284d841 (from Discourse thread above) I get CPU:
```
julia> main()  # 2nd run:
181.671379 seconds (273.90 k allocations: 49.164 GiB, 90.20% gc time)  # before
141.337037 seconds (257.10 k allocations: 34.292 GiB, 87.77% gc time)  # after

julia> main()  # on a different computer! Julia 1.8, first run
 46.989183 seconds (2.30 M allocations: 49.275 GiB, 36.97% gc time, 11.01% compilation time)  # before
 35.523793 seconds (2.27 M allocations: 34.402 GiB, 14.94% gc time, 14.46% compilation time)  # after
```
GPU, inserting `|> gpu` and `CUDA.@time`. Quite noisy, best of a few each, same device:
```
julia> main()
# before
  1.994243 seconds (2.26 M CPU allocations: 213.138 MiB, 7.88% gc time) (61.50 k GPU allocations: 49.165 GiB, 36.31% memmgmt time)
# after
  1.800311 seconds (2.20 M CPU allocations: 213.963 MiB, 4.43% gc time) (53.70 k GPU allocations: 34.287 GiB, 17.56% memmgmt time)

julia> CUDA.device()
CuDevice(3): Tesla P100-PCIE-16GB
```